requirements.txt
gunicorn==19.9.0
Werkzeug==0.15.5
pandas>=0.19
asttokens==2.0.5
certifi==2021.5.30       
charset-normalizer==2.0.2
click==8.0.1
colorama==0.4.4
cycler==0.10.0
executing==0.7.0
Flask==2.0.1
icecream==2.1.1
idna==3.2
importlib-metadata==4.6.1
itsdangerous==2.0.1      
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.0
pandas==1.3.0
Pillow==8.3.1
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.0
selenium==3.141.0
six==1.16.0
threadpoolctl==2.2.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0